Change CursorLockMode to Locked instead of Confined
With the CursorLockMode set to confined, at least on Linux, the cursor was able to move to the edges of the window at which point it would no longer be possible to rotate the camera further in that direction. Changing it to locked means that the cursor is locked in the center of the window, which fixes this issue.